Organized a few TODOs.
git-svn-id: https://s.snth.net/svn/rlk/thumb@974 dcf3a027-e0e4-0310-89c9-ab9954db6d00

+++ src/batcher.cpp
@@ -176,7 +176,7 @@ GLuint *ogl::segment::reduce(vert_p& v, GLuint *e,
 
     for (mesh_elem_i i = meshes.begin(); i != meshes.end(); ++i)
     {
-        GLuint min, max, d = GLuint(v) / sizeof (vert);
+        GLuint min, max, d = GLuint(size_t(v)) / sizeof (vert);
 
         // Cache the vertices to the vertex buffer.
